ENH: Change how we fall back.
Only write data in a fallback location on S3 if the data we have are
smaller than what's there. Otherwise, overwrite. This should help the
condition where we get some new chunks much later (after the volume
originally arrived) and we pull the chunks from cache and assemble a
more complete volume.