ENH: Tweak logic for completing volume.
Make checking if the volume is done its own top-level method. Check this
on every iteration of the while loop, instead of having a flag returned
from add. What this handles is the case where the volume loading from
cache (which happens while waiting for it to be ready) ends up
completing the volume. This would happen if any chunk shows up that
completes the volume with those in the cache. Without this fix, we would
need to wait for the volume to timeout.